requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.74
botocore==1.34.74
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-environ==0.11.2
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
jmespath==1.0.1
kombu==5.3.6
mysqlclient==2.2.4
pillow==10.2.0
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.9.0.post0
redis==5.0.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
